requirements.txt
fastapi~=0.101.0
uvicorn
pydantic~=2.11.7
openai~=1.99.6
python-dotenv~=1.0.0